server/pxf-service/src/main/java/org/greenplum/pxf/service/UGICache.java
package org.greenplum.pxf.service;

/*
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

import com.google.common.base.Ticker;
import org.apache.hadoop.security.UserGroupInformation;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.stereotype.Component;

import java.io.IOException;
import java.util.HashMap;
import java.util.Map;
import java.util.concurrent.ConcurrentHashMap;
import java.util.concurrent.DelayQueue;
import java.util.concurrent.Delayed;
import java.util.concurrent.TimeUnit;
import java.util.concurrent.atomic.AtomicInteger;

/**
 * Stores UserGroupInformation instances for each active session. The UGIs are cleaned up if they
 * have not been accessed for UGI_CACHE_EXPIRY milliseconds.
 * <p>
 * The motivation for caching is that destroying UGIs is slow. The alternative, creating and
 * destroying a UGI per-request, is wasteful.
 */
@Component
@SuppressWarnings("UnstableApiUsage")
public class UGICache {

    static final int NANOS_PER_MILLIS = 1000000;
    static final long UGI_CACHE_EXPIRY = 15 * 60 * 1000L; // 15 Minutes
    private static final Logger LOG = LoggerFactory.getLogger(UGICache.class);
    private final Map<SessionId, Entry> cache = new ConcurrentHashMap<>();
    // There is a separate DelayQueue for each segment (also being used for locking)
    private final Map<Integer, DelayQueue<Entry>> expirationQueueMap = new HashMap<>();
    private final UGIProvider ugiProvider;
    private final Ticker ticker;

    /**
     * Create a UGICache with the given {@link Ticker} and {@link UGIProvider}. Intended for use by
     * tests which need to mock UGI creation/destruction and the current time.
     */
    UGICache(UGIProvider provider, Ticker ticker) {
        this.ticker = ticker;
        this.ugiProvider = provider;
    }

    /**
     * Create a UGICache. Automatically creates a {@link UGIProvider} that this cache will use to
     * create and destroy UserGroupInformation instances.
     */
    public UGICache() {
        this(new UGIProvider(), Ticker.systemTicker());
    }

    /**
     * If a UGI for the given session exists in the cache, returns it. Otherwise, creates a new
     * proxy UGI. In either case this method increments the reference count of the UGI. This method
     * also destroys expired, unreferenced UGIs for the same segmentId as the given session.
     *
     * @param session     The user from the session is impersonated by the proxy UGI.
     * @param isProxyUser true if the {@link UserGroupInformation} is a proxy user
     * @return the proxy UGI for the given session.
     * @throws IOException when there is an IO issue
     */
    @SuppressWarnings("SynchronizationOnLocalVariableOrMethodParameter")
    public UserGroupInformation getUserGroupInformation(SessionId session, boolean isProxyUser) throws IOException {
        Integer segmentId = session.getSegmentId();
        String user = session.getUser();
        DelayQueue<Entry> delayQueue = getExpirationQueue(segmentId);
        synchronized (delayQueue) {
            // Use the opportunity to cleanup any expired entries
            cleanup(delayQueue);
            Entry entry = cache.get(session);
            if (entry == null) {

                UserGroupInformation ugi;
                if (isProxyUser) {
                    LOG.debug("{} Creating proxy user = {}", session, user);
                    ugi = ugiProvider.createProxyUGI(user, session.getLoginUser());
                } else {
                    LOG.debug("{} Creating remote user = {}", session, user);
                    ugi = ugiProvider.createRemoteUser(user, session);
                }
                entry = new Entry(ticker, ugi, session);
                delayQueue.offer(entry);
                cache.put(session, entry);
            }
            entry.incrementRefCount();
            return entry.getUGI();
        }
    }

    /**
     * Decrement reference count for the given session's UGI. Resets the time at which the UGI will
     * expire to UGI_CACHE_EXPIRY milliseconds in the future.
     *
     * @param session                  the session for which we want to release the UGI.
     * @param cleanImmediatelyIfNoRefs if true, destroys the UGI for the given session (only if it
     *                                 is now unreferenced).
     */
    @SuppressWarnings("SynchronizationOnLocalVariableOrMethodParameter")
    public void release(SessionId session, boolean cleanImmediatelyIfNoRefs) {

        Entry entry = cache.get(session);

        if (entry == null) {
            throw new IllegalStateException("Cannot release UGI for this session; it is not cached: " + session);
        }

        DelayQueue<Entry> expirationQueue = getExpirationQueue(session.getSegmentId());

        synchronized (expirationQueue) {
            entry.decrementRefCount();
            expirationQueue.remove(entry);
            if (cleanImmediatelyIfNoRefs && entry.isNotInUse()) {
                closeUGI(entry);
            } else {
                // Reset expiration time and put it back in the queue
                // only when we don't close the UGI
                entry.resetTime();
                expirationQueue.offer(entry);
            }
        }
    }

    /**
     * @return the size of the cache
     */
    int size() {
        return cache.size();
    }

    /**
     * This method is not thread-safe, and is intended to be called in tests.
     *
     * @return the sum of the sizes of the internal queues
     */
    int allQueuesSize() {
        int count = 0;
        for (DelayQueue<UGICache.Entry> queue : expirationQueueMap.values()) {
            count += queue.size();
        }
        return count;
    }

    /**
     * This method is O(n) in the number of cache entries and should only be called in tests.
     *
     * @param session the session
     * @return determine whether the session is in the internal cache
     */
    @SuppressWarnings("SynchronizationOnLocalVariableOrMethodParameter")
    boolean contains(SessionId session) {
        DelayQueue<Entry> expirationQueue = getExpirationQueue(session.getSegmentId());
        synchronized (expirationQueue) {
            Entry entry = cache.get(session);
            return entry != null && expirationQueue.contains(entry);
        }
    }

    /**
     * Get the queue of cache entries associated with a segment, creating it if it doesn't yet
     * exist. This lets us lazily populate the expirationQueueMap.
     *
     * @param segmentId the segment identifier
     * @return the {@link DelayQueue} associated to the segment.
     */
    private DelayQueue<Entry> getExpirationQueue(Integer segmentId) {
        DelayQueue<Entry> queue = expirationQueueMap.get(segmentId);
        if (queue == null) {
            synchronized (expirationQueueMap) {
                queue = expirationQueueMap.get(segmentId);
                if (queue == null) {
                    queue = new DelayQueue<>();
                    expirationQueueMap.put(segmentId, queue);
                }
            }
        }
        return queue;
    }

    /**
     * Iterate through all the entries in the queue and close expired {@link UserGroupInformation},
     * otherwise it resets the timer for every non-expired entry.
     *
     * @param expirationQueue the expiration queue
     */
    private void cleanup(DelayQueue<Entry> expirationQueue) {

        Entry expiredUGI;
        while ((expiredUGI = expirationQueue.poll()) != null) {
            if (expiredUGI.isNotInUse()) {
                closeUGI(expiredUGI);
            } else {
                // The UGI object is still being used by another thread
                String fsMsg = "FileSystem for proxy user = " + expiredUGI.getSession().getUser();
                LOG.debug("{} Skipping close of {}", expiredUGI.getSession().toString(), fsMsg);
                // Place it back in the queue if still in use and was not closed
                expiredUGI.resetTime();
                expirationQueue.offer(expiredUGI);
            }
            LOG.debug("Delay Queue Size for segment {} = {}", expiredUGI.getSession().getSegmentId(), expirationQueue.size());
        }
    }

    /**
     * This method must be called from a synchronized block for the delayQueue for the given
     * session.getSegmentId(). Removes the cachedUGI from the internal cache and then passes it to
     * {@link UGIProvider} to destroy the UGI.
     *
     * @param expiredUGI the UserGroupInformation to close
     */
    private void closeUGI(Entry expiredUGI) {
        SessionId session = expiredUGI.getSession();
        String fsMsg = "FileSystem for proxy user = " + session.getUser();

        try {
            // Remove it from cache, as cache now has an
            // expired entry which is not in progress
            cache.remove(session);
            ugiProvider.destroy(expiredUGI.getUGI());

        } catch (Throwable t) {
            LOG.warn(session.toString() + " Error closing " + fsMsg, t);
        }

        LOG.debug("{} Closed {} (Cache Size = {})", session.toString(), fsMsg, cache.size());
    }

    /**
     * Stores a {@link UserGroupInformation}, and determines when to expire the UGI.
     */
    @SuppressWarnings("UnstableApiUsage")
    private static class Entry implements Delayed {

        private final SessionId session;
        private final UserGroupInformation proxyUGI;
        private final AtomicInteger referenceCount = new AtomicInteger();
        private final Ticker ticker;
        private volatile long startTime;

        /**
         * Creates a new UGICache Entry.
         *
         * @param ticker   the time ticker
         * @param proxyUGI the proxy UserGroupInformation
         * @param session  the session
         */
        Entry(Ticker ticker, UserGroupInformation proxyUGI, SessionId session) {
            this.ticker = ticker;
            this.proxyUGI = proxyUGI;
            this.session = session;
        }

        /**
         * @return the Cached {@link UserGroupInformation}.
         */
        public UserGroupInformation getUGI() {
            return proxyUGI;
        }


        /**
         * @return the session associated to the {@link UserGroupInformation}.
         */
        public SessionId getSession() {
            return session;
        }

        /**
         * @return true if the UGI is being referenced by a session, false otherwise
         */
        private boolean isNotInUse() {
            return referenceCount.get() <= 0;
        }

        /**
         * Increments the number of references accessing the {@link UserGroupInformation}.
         */
        void incrementRefCount() {
            referenceCount.incrementAndGet();
        }

        /**
         * Decrements the number of references accessing the {@link UserGroupInformation}.
         */
        void decrementRefCount() {
            int count = referenceCount.decrementAndGet();
            if (count < 0) {
                throw new IllegalStateException("UGICache.Entry referenceCount may not be decremented past 0.");
            }
        }

        /**
         * Resets the timer for removing this Entry from the cache.
         */
        void resetTime() {
            startTime = currentTimeMillis();
        }

        /**
         * {@inheritDoc}
         */
        @Override
        public long getDelay(TimeUnit unit) {
            return unit.convert(getDelayMillis(), TimeUnit.MILLISECONDS);
        }

        /**
         * Compare the expiry time of this cache entry to another cache entry's expiry time.
         *
         * @param other a UGICache.Entry (passing any other kind of Delayed produces an error)
         * @see java.lang.Comparable<>#compareTo(java.lang.Comparable<>)
         */
        @Override
        public int compareTo(Delayed other) {
            if (!(other instanceof Entry)) return 1;

            Entry that = (Entry) other;
            return Long.compare(this.getDelayMillis(), that.getDelayMillis());
        }

        /**
         * @return the number of milliseconds remaining before this cache entry expires.
         */
        private long getDelayMillis() {
            return (startTime + UGI_CACHE_EXPIRY) - currentTimeMillis();
        }

        /**
         * @return the current Unix timestamp in milliseconds (equivalent to {@link
         * System}.currentTimeMillis)
         */
        private long currentTimeMillis() {
            return ticker.read() / NANOS_PER_MILLIS;
        }
    }
}
